+++ eth2util/hash_test.go
@@ -1,0 +1,34 @@
+// Copyright © 2022 Obol Labs Inc.
+//
+// This program is free software: you can redistribute it and/or modify it
+// under the terms of the GNU General Public License as published by the Free
+// Software Foundation, either version 3 of the License, or (at your option)
+// any later version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of  MERCHANTABILITY or
+// FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+// more details.
+//
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <http://www.gnu.org/licenses/>.
+
+package eth2util_test
+
+import (
+	"encoding/hex"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+
+	"github.com/obolnetwork/charon/eth2util"
+)
+
+func TestSlotHashRoot(t *testing.T) {
+	resp, err := eth2util.SlotHashRoot(2)
+	require.NoError(t, err)
+	require.Equal(t,
+		"0200000000000000000000000000000000000000000000000000000000000000",
+		hex.EncodeToString(resp[:]),
+	)
+}
